+++ libraries/transformer/src/test/java/androidx/media3/transformer/EditedMediaItemBuilderTest.java
@@ -232,43 +232,6 @@ public final class EditedMediaItemBuilderTest {
   }
 
   @Test
-  public void setSpeed_withSetSpeedChangingEffects_doesNotThrow() {
-    TimestampAdjustment timestampAdjustment =
-        new TimestampAdjustment((inputTimeUs, outputTimeConsumer) -> {}, SPEED_PROVIDER_2X);
-    SpeedChangingAudioProcessor processor = new SpeedChangingAudioProcessor(SPEED_PROVIDER_2X);
-    EditedMediaItem unused =
-        new EditedMediaItem.Builder(MediaItem.EMPTY)
-            .setSpeed(SPEED_PROVIDER_2X)
-            .setSpeedChangingEffects(processor, timestampAdjustment)
-            .build();
-  }
-
-  @Test
-  public void setSpeed_withSetSpeedChangingEffects_throwsWithMismatchingSpeedProviders() {
-    SpeedChangingAudioProcessor processor = new SpeedChangingAudioProcessor(SpeedProvider.DEFAULT);
-    assertThrows(
-        IllegalStateException.class,
-        () ->
-            new EditedMediaItem.Builder(MediaItem.EMPTY)
-                .setSpeed(SPEED_PROVIDER_2X)
-                .setSpeedChangingEffects(processor, /* effect= */ null)
-                .build());
-  }
-
-  @Test
-  public void setSpeedChangingEffects_withMismatchingSpeedProviders_throws() {
-    TimestampAdjustment timestampAdjustment =
-        new TimestampAdjustment((inputTimeUs, outputTimeConsumer) -> {}, SPEED_PROVIDER_2X);
-    SpeedChangingAudioProcessor processor = new SpeedChangingAudioProcessor(SpeedProvider.DEFAULT);
-    assertThrows(
-        IllegalArgumentException.class,
-        () ->
-            new EditedMediaItem.Builder(MediaItem.EMPTY)
-                .setSpeedChangingEffects(processor, timestampAdjustment)
-                .build());
-  }
-
-  @Test
   public void setPreProcessingAudioProcessors_populatesPreProcessingAudioProcessors() {
     ToInt16PcmAudioProcessor processor = new ToInt16PcmAudioProcessor();
     EditedMediaItem item =
@@ -281,13 +244,6 @@ public final class EditedMediaItemBuilderTest {
     EditedMediaItem duplicateItem = item.buildUpon().build();
 
     assertThat(duplicateItem.preProcessingAudioProcessors).containsExactly(processor);
-  }
-
-  @Test
-  public void getDurationAfterEffectsApplied_withSpeedProvider_returnsCorrectDuration() {
-    EditedMediaItem editedMediaItem =
-        new EditedMediaItem.Builder(MediaItem.EMPTY).setSpeed(SPEED_PROVIDER_2X).build();
-    assertThat(editedMediaItem.getDurationAfterEffectsApplied(1_000_000)).isEqualTo(500_000);
   }
 
   @Test